Added simple mutex class based on semaphore to guard critical secions; it is for example used in the process where the log file is written in the logbuf class

+++ src/util/sigsegv.c
@@ -59,7 +59,7 @@ static void signal_segv(int signum, siginfo_t* info, void*ptr) {
     fprintf(stderr, "info.si_code  = %d (%s)\n", info->si_code, si_codes[info->si_code]);                         
     fprintf(stderr, "info.si_addr  = %p\n", info->si_addr);                                                       
     for(i = 0; i < NGREG; i++)                                                                                    
-        fprintf(stderr, "reg[%02d]       = 0x" REGFORMAT "\n", i, ucontext->uc_mcontext.gregs[i]);                
+        fprintf(stderr, "reg[%02d]       = 0x" REGFORMAT "\n", (int)i, ucontext->uc_mcontext.gregs[i]);                
                                                                                                                   
 #if defined(SIGSEGV_STACK_X86) || defined(SIGSEGV_STACK_IA64)                                                     
 # if defined(SIGSEGV_STACK_IA64)
